+++ CHANGELOG.md
@@ -10,6 +10,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added stderr to output of ShellExecute functions
 - Added ShellCmdExecute function
 
+### Changed
+- Updated mimikatz binaries
+- Changed mimikatz function to load in new thread, free input/output pointers
+
 ### Fixed
 - Fixed XML warning, removed angle brackets in comment